+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,0 +1,16 @@
+package bowling.domain.frame;
+
+import bowling.domain.score.Score;
+
+public interface Frame {
+    public static final int MAXIMUM_NORMAL_PITCH_COUNTS = 2;
+    public static final int MAXIMUM_FINAL_PITCH_COUNTS = 3;
+
+    public Frame next();
+
+    public void bowl(Score score);
+
+    public boolean isMovableToNextFrame();
+
+    public int getIndex();
+}
