+++ src/main/java/bowling/domain/pitch/Pitches.java
@@ -1,0 +1,12 @@
+package bowling.domain.pitch;
+
+import bowling.domain.score.Score;
+
+import java.util.List;
+
+public interface Pitches {
+
+    public void throwBall(Score score);
+
+    public List<Pitch> getPitches();
+}
